flash (flashrom script): remove boardmismatch=force
This was put there before for users upgrading from libreboot r5
to r6, but also allows the user to flash the wrong image. For
example, the user could flash a T60 image on an X60, thus
bricking the machine. It's almost certain that most people
have upgraded by now, so remove this potentially dangerous
option.

If a user is upgrading from r5 to the latest version now,
then they can simply use flashrom directly.