buildrom-withgrub: Only include 1 keymap file in GRUB memdisk.
This is to reduce the size of the image. To accomplish this,
the payload in each coreboot configuration was set to 'None',
after which the buildrom-withgrub script was modified to
build a new grub.elf for each image. This is fast, since
building the grub.elf only takes less than a second. The
alternative was to build a new coreboot ROM per keymap,
which would have been slow.

Other files were also modified along with this, check
the diff to see everything.