build: fix bug: images for i945 ROM's were not dd'd
For X60/X60T/T60

An optimization was made earlier to reduce the number of steps
taken but it was found that the ROM images did not contain
identical data in the top two 64K regions of the image.

This would have bricked the users machine if the followed the
correct instructions, even. This patch fixed the issue.