grub.cfg: hardcode the list of partitions to search
For libreboot_grub.cfg or ISOLINUX parsing, the old grub.cfg was
using regexp to automatically search all partitions. For example,
it was using (ahci0,*). This, and the ls, is extremely slow on
some machines. Partition/file access is itself fast, but producing
the list can be slow.

Therefore, use hardcoded entries like ahci0,1-4 (up to 4) or
usb0,1-4 instead of *. This is less flexible, but much faster
and it should still work for most people.