buildrom-withgrub: Put background.jpg in CBFS, not GRUB memdisk
The background.jpg was put in CBFS originally to save space inside
CBFS because the GRUB payload is LZMA compressed by coreboot. The
actual measured saving in space for background.jpg being in the
memdisk was about 4KiB.

Putting it in CBFS means that the user can easily change their
background image if they want to.