docs/tasks.html: remove X201 from candidate list (waste of time)
Any Intel-based laptop using a chipset newer than GM45 is a waste
of time, because the ME is too difficult to remove/replace and
would likely mean that the system no longer works properly; for
example, power management will probably stop working, which is
bad for laptops generally.

The ME is signed, so replacing it simply isn't feasible.