External flashing guides: don't refer to red.
Colour blindness is real, and quite common.

What's worse, 3.3V on an ATX PSU is an orange wire, not a red
wire. Someone looking at "red" in the documentation might not
think that you mean 3.3V live/hot (red is the standard colour),
but instead they might connect the red wire from the ATX PSU
which is 5V; this will destroy the flash chip. Also, red and
orange look the same to most colour blind people.