disable generation of txtmode ROM images (temporarily)
I wanted to get a release done as quickly as possible (within 24
hours after pushing this commit). That release will most likely
be short-lived; as soon as it's out, the plan is to work as fast
as possible fixing as many bugs as possible (there are a lot)
and make a new release again as soon as possible.