Update coreboot again (a patch was merged upstream)
The patch for only requiring cmake for clang users was merged.
This patch is important, because libreboot doesn't want to use
clang, and doesn't want any dependences that it relies on which
it doesn't need.

Also, this and the other recent update re-add support for ACPI
brightness methods on the Thinkpad X60 and T60.