docs: Make the C201 hcl page more factual (with corrections), without ordering users
This rewrites parts of the C201 hcl page in more factual ways. This adds details
about the relationship between Google's Chromebook-related decisions and free
software, without giving a too strong opinion about Google given that it's a
company so big that it makes little sense to judge it as a whole.

Also, a few sentences were rewritten to not tell users what to do or what to
think, but to inform them about the current problems and solutions when it comes
to running the device with free software.

Signed-off-by: Paul Kocialkowski <contact@paulk.fr>