grub.cfg: remove option for isolinux parsing on CD/DVD
The option was misleading, since only CDs work. Optic media is
obsolete, and we recommend using USB instead. Most libreboot
systems don't even have optic media drives to begin with. Not only
that, but there was no real guarantee that an optical disc drive
would be present on ata0/ahci1 anyway, and was just an approximated
assumption/guess.