Download, compile and clean crossgcc separately to coreboot
This means that coreboot can be re-downloaded and re-built without affecting
crossgcc. The crossgcc directory also now only contains crossgcc; previously,
it contained the entire coreboot source code too, which was unneeded because
that directory was just for crossgcc, where the real coreboot directories
symlink to it during the build process.