build/release/src: don't delete patches for coreboot
The old model was to have 1 coreboot directory with 1 revision, with all patches
for all boards, so this behaviour used to be correct. Now it is wrong, because
different coreboot revisions are used in different directories, and the build
process does git-init and applies those patches. So we need them.