set vram on gm45 to 352M
This patch enables all vram sizes and sets the default to 352M for
libreboot.
Normally users have to clear their vram for this to happen. using
"nvram -c random"

But there a 2 better option:
coreboot provides to always load the defaults in its menuconfig
update the flashscript to reset cmos

Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>