Add boards: use a static CMOS option table, for consistency
This also forces GM45 to use 352MiB RAM for the Video RAM, for the Intel GPU.
It's implemented in such a way where the user configures it using nvramtool.

Hardcode it.

This should also force the century byte to be reset on GM45, for those users
who were affected by this previous issue.