Cleanup/tidy the texinfo documentation, courtesy of Paul <paukoen@gmail.com>
This was submitted on the mailing list. I further added the full list of
authors to the copying section. There were a few copyright authors missing in
the file that Paul submitted.

There are some TODO entries at the start of the document, which Paul left for
us. We need to work on those.

Also TODO: integrate this into build system, so that we have these options:
./build docs html
./build docs html-onepage
./build docs pdf
etc