Add sound card config for x60
The sound card on the x60 is not configured in coreboot.
This causes the headphones not to be automatically
detected/used when inserted which it does on the vendor
bios. The internal microphone is also not set as default.

I found on the coreboot wiki how to get the config from vendor bios.
In GNU/linux run:
 for x in /sys/class/sound/card0/hw*; do cat "$x/init_pin_configs" > pin_"$(basename "$x")"; done
 for x in /proc/asound/card0/codec#*; do cat "$x" > "$(basename "$x")"; done
which gives the pin config of the vendor bios.