Re-add suppliers page, with all known suppliers listed
Sorted into alphabetical order, to avoid bias. Each product listed, for each
supplier.

This page needs to be re-added, so that people know where to get
libreboot-preinstalled hardware. Not everyone is capable of installing
Libreboot on their own, and our previous decision to not have a suppliers page
is hurting Libreboot's ability to gain additional users.