Add Andrew Robbins to core management team (with permission from other leaders)
Permission was granted by Paul, Alyssa and Swift in a private email thread.

Andrew is working on huge changes to the build system to make it usable, and
understands most of Libreboot and is therefore suitable to review and merge
patches that are contributed to the project.

Andrew has also demonstrated an understanding in how the documentation is
maintained, and generally seems to understand Libreboot's practises well and
agrees with them. Andrew is a perfect fit for the core management team, in my
opinion, and I look forward greatly to their future work within Libreboot.