Improved handling of filenames in archive creation
Filenames listed in DOTTARFILES are now NUL terminated so that
any with strange characters, such as a newline, do not negatively
affect archive creation.  DOTRNDSEED file which stores the random seed
to be used by GCC during compilation is now included in the list of
files to be archived.  Cleaned up functions related to archive creation.