Fixed printf calls & replaced 'echo' w/ printf.
All printf calls should now be properly formatted; prior, the
format specifier string was erroneously used for both the format
specifiers and the string to be printed.  'env' is now used to
locate the printf binary so as to avoid potentially using a shell
builtin.  Lastly, all calls to 'echo' within the new build system
have been replaced with printf for consistency/portability purposes.