docs/publish.sh: Fix sed(1) invocation on OpenBSD
On OpenBSD, publish.sh errors out at the following call
to sed(1) (from "bash -x publish.sh"):

    + sed temp.md -i -e 's/\.md\(#[a-z\-]*\)*)/.html\1)/g'
    sed: 1: "temp.md": undefined label 'emp.md'

It seems that "temp.md" is being parsed by sed as a sed command,
not as a named file. This is likely due to OpenBSD's strict
usage requirements for sed:

    usage: sed [-aEnru] [-i[extension]] command [file ...]
           sed [-aEnru] [-e command] [-f command_file] [-i[extension]] [file ...]

As shown above, the sed command must always come before any named
files. This commit does that, which fixes publish.sh with OpenBSD's sed.
This is also tested and working with GNU sed v4.2.2