Fix action text in './libreboot usage <project>'
Apparently the name used in a 'for /name/ ...' loop rebinds any local
variables (with the same name) with whatever /name/ is bound
to in a way that persists even after the loop returns.  The crux of
the issue here is that a function's children can rebind the parent(s)'
local variables just by using the same name as the variable in a for
loop, which is surprising--and apparently undocumented--behavior.

Use of a subshell group for encapsulating the for loops (See:
project_usage_actions) mitigates the aforementioned issue.

Closes issue: #244