www/publish.sh: Fix conditionals for file matching
Before the Makefile, publish.sh was executed on markdown source
files using find(1), which happened like this:

    ./publish.sh ./index.md

Now that we have a Makefile, this happens instead:

    ./publish.sh index.md

Note that the file argument "index.md" in the first and second case
both refer to the same file, yet they are different strings. This is
important because publish.sh gives index.md (among other files) special
treatment, and it does this by string comparison.

Unfortunately, only the argument in the first case ("./index.md") will
cause publish.sh to give special treatment, while the argumnent in the
second case ("index.md") will not.

To fix this, make it so that both "./index.md" and "index.md" trigger
publish.sh's special handling. This commit also fixes the same issue
for "docs/fdl-1.3.md" and "conduct.md".