www/publish.sh: Fix quoting bug
When running "bash -x publish.sh index.md", the following happens:

    + pandoc --smart /tmp/libreboot_www.Xne7SYgf4e -s --css /global.css '--css /headercenter.css' --template template.html --metadata return=
    pandoc: unrecognized option `--css /headercenter.css'
    Try pandoc --help for more information.

This is because the OPTS variable is being quoted, which causes its
value of "-css /headercenter.css" to not be split on word boundaries,
that is, spaces. Because pandoc has a "--css" option, but not a
"--css /headercenter.css" option, pandoc expectedly complains that
said option does not exist.

To fix this we just unquote OPTS, and add a shellcheck ignore directive
so this won't be reintroduced in the future.

Caused by: a8d89665de6da20a8793886e03f153e922f6e519.