remove minifree d16 from suppliers (pull request not needed for this commit)
this concerns an entry on the suppliers page, for a product from my own
company. Pull request therefore not required

I've decided to take D16 off of minifree, in order to reduce competition for
the TALOS II product that Raptor Engineering launched today. The world needs
libre OEMs, and Raptor Engineering is the first one.