Consolidate GRUB BIOS target modules into 'install'
Since the majority of GRUB modules will be placed into CBFS instead of
the GRUB image itself, the modules-install and modules-preload files
no longer have any use as-is.  However, they serve as a useful
reference to which modules should be retained for later
processing (thus the addition to the new install file).