Create GRUB config files font-file, font-project
'font-file' contains the filename (not path) of the font to use when
making a PF2-format GRUB font.

'font-project' contains the name of the project which built the
original font file.

These files are necessary to avoid hard-coding the font and/or the
path to the font as an argument to grub-mkfont.