Remove overlooked braces from parameter expansions
This is a continuation of the last set of commits removing braces from
parameter expressions which do not require them.  The main script was
overlooked when applying the previous changes, which this commit aims
to correct.

Also, one parameter expansion in libs/common was corrected as it was
overlooked as well.