Conditionally bootstrap CrossGCC build
It's possible that a compilation failure will occur if there's a
difference in major versions between the host GCC compiler and the one
being built.  To avoid this, bootstrapping can be used.  The method
for bootstrapping is simply passing the '-b' flag to Make; the
Makefile takes care of the rest.