Replace usage of the '[' Bash builtin with '[['
There's no benefit to using the POSIX-style '[' test builtin
considering its '-a' and '-o' operators are unused in the Libreboot
build system.  Plus, '[[' is safer with respect to any containing file
redirections (for example).

Prior, both '[' and '[[' were used throughout the codebase--a
disparity in usage which this change aims to eliminate.