Remove unnecessary spaces in command substitutions
This is likely one of the very last changes necessary to make the
Libreboot build system more cohesive in appearance.

Hopefully from this point forward it won't be as readily apparent as
to who wrote which parts of the build system (i.e. won't look like a
patchwork quilt any longer).