Use project_sources_path() in archive creation
Previously, the function 'project_release_sources_archive_path'
assumed that a project's source files would always be located within
"$root/$SOURCES" in a directory named after the project.  However,
this does not account for projects such as bucts whose source files
are located in "$root/$PROJECTS/$project/sources".

In order to address this, trivial changes were made to leverage
project_sources_path() as it returns nearly all the information needed
to create the path to the archive file.