Update download page to include the 16-digit GPG key (no pull request needed)
This was brought to our attention by attila_lendvai on #libreboot IRC.
Essentially, the previous instruction was vulnerable to potential collision
attacks.

With this new instruction, it should be a bit safer.

Since this is urgent, no pull request was issued. This has been committed to
the repository directly, with consent from and_who on IRC.