Add temporary link to all pages on the website, for Leah's surgery fundraiser
This isn't technically libreboot-related, but I'm currently running a
fundraiser to try to get donations, so that I can pay for my Gender
Reassignment Surgery. This was pushed to Pull Review, subject to review by
other members of core management before being published on libreboot.org

I modified the template.html document, so that this appears on all pages.