Pending change for CMOS value naming 'switchable graphics' -> 'dual graphics'
I proposed a naming patch for the T400 CMOS value in the Coreboot project: https://review.coreboot.org/#/c/coreboot/+/27158/
As this is the only manual referring to the name of the value, this commit updates the naming to reflect the upstream change.

Merging this patch becomes relevant once the coreboot patch has made it's way into Libreboot.