Add screen, stylus and trackpoint rotation script
This script's purpose is to rotate the screen and/or Wacom stylus
and/or trackpoint orientation for more comfortable drawing and writing
on a touch screen. Requires `xrandr` and `xsetwacom` on the system.

Various desktop environments have different levels of automation, so
options `-s`, `-p`, `-t` determine whether to rotate the screen, pen
or trackpoint.

Only the main screen is rotated (LVDS1 or LVDS-1 depending on
kernel). Furthermore the input device rotations follow the main
screen.

_THERE MIGHT BE ISSUES WITH EXTERNAL MONITORS_ Check!