Create 4MiB Coreboot config/target for Veyron Mickey
4MiB flash is the default for this board. A 16MiB config is not
included (yet) as I was unable to find a teardown of this device
(Asus Chromebit CS10) online to be sure that reassembly is possible.

Also, corebootfb is the only display mode available for this board;
as such, inclusion of the corebootfb subtarget serves only to
explicitly indicate the display mode when packaging ROMs upon
release.

Veyron Mickey Coreboot ROMs can be built with, e.g.:
'./libreboot build coreboot veyron mickey corebootfb 4mb'