Rework regexp for .md extension replacement (www)
Related issue #570

The markdown file extension was not consistently replaced with an
HTML extension due to a bug in the regexps used in www/publish.sh
caused by the A-Z character range not being included in the
character alternative.

Along with correcting the aforementioned, this patch allows for more
robust link anchor capturing by allowing the use and replacement of
characters within the ranges a-z, A-Z, 0-9, _ (underscore),
and - (hyphen).