Create 8MiB target for qemu_q35_ich9
The 16MiB ROM is non-functional at the moment so an 8MiB target has
been added in order to avoid losing QEMU support while the bug is
identified and fixed.

Files were copied from the 16MiB directory with the exception of the
config file located at 8MiB/config which is new.