libs/git: Create branch name based on argument list
There was a bug where only $project was used as the branch name due to
some faulty logic which checked for the existence of a revision file
before setting the variables "prepare_branch" and "prepare_path".

The bug in this case was that prepare_branch would be set to the value
of $branch in only the first iteration of the for loop since in almost
all cases there is only one revision file that exists for a given
project, regardless of the arguments provided. Explained another way,
in order for the proper branch name to be used a revision file would
have had to exist in every target directory.

This was an issue because only one branch was ever operated on (named
$project), meaning if actions were performed on many targets then only
the last to run would be represented in the project's
repository--making tracking down some bugs a bit harder.

With this fixed we now create a branch for every possible project
or tool configuration and leaves us with a log for each.