libs/git: Add git_apply() and git_apply_check()
Since changes from an applied diff need to be added to the index and
committed (vs. mailbox files with the git_patch function) new
functions were necessary to account for this difference.

This also fixes a bug where diffs were applied but not committed,
at minimum dirtying the work tree and potentially causing problems for
any later patches.