libs/project: Print project_action() footer, always
Previously, the footer would not print if the project action failed
due to 'set -e' still being in effect (the subshell would return with
a non-zero exit code and exit immediately). The purpose of the
subshell for the project action is to better separate the execution
environment of the project from the library functions it
calls.

The test condition in the if-statement and its consequents were
changed in order to be more concise and informative.