Update BBB configuration for Debian 9
SSH access to a BBB has changed as of Debian 9, with the user needing to
login to the debian account instead of root. Furthermore, the addition
of the config-pin utility has simplified spidev setup considerably,
with the device tree overlay files no longer being needed. References to
the BBB screwdriver have also been removed, since it appears to no
longer be maintained.